backend/requirements.txt
# FastAPI et dépendances
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Base de données
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentification et sécurité
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Redis
redis==5.0.1
hiredis==2.3.2

# Celery pour tâches asynchrones
celery==5.3.6

# HTTP et CORS
httpx==0.26.0
python-dotenv==1.0.0

# Monitoring et logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Validation et utilitaires
email-validator==2.1.0
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
faker==22.1.0

# Code quality
black==24.1.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0

# Development
ipython==8.20.0